+++ node/directives/checkUserAccess.ts
@@ -24,7 +24,17 @@ export async function checkUserOrAdminTokenAccess(
 
   if (adminUserAuthToken) {
     try {
-      await identity.validateToken({ token: adminUserAuthToken })
+      const authUser = await identity.validateToken({
+        token: adminUserAuthToken,
+      })
+
+      if (!authUser?.audience || authUser?.audience !== 'admin') {
+        logger.warn({
+          message: `CheckUserAccess: No valid user found by admin token`,
+          operation,
+        })
+        throw new ForbiddenError('Unauthorized Access')
+      }
     } catch (err) {
       logger.warn({
         error: err,